Set up basic structure of writing helm
