+++ README.md
@@ -37,6 +37,7 @@ It is not possible to run a full Taskcluster deployment directly from this repos
 * [Infrastructure](infrastructure#readme)
     * [Taskcluster Builder](infrastructure/builder#readme)
     * [Test Docker Image](infrastructure/docker-images#readme)
+    * [infrastructure/helm](infrastructure/helm#readme)
     * [References](infrastructure/references#readme)
     * [Terraform](infrastructure/terraform#readme)
 * [Libraries](libraries#readme)
