+++ infrastructure/helm/templates/NOTES.txt
@@ -1,0 +1,1 @@
+Big Mistake! Should not have used Taskcluster!
