+++ infrastructure/helm/templates/auth.yaml
@@ -1,0 +1,4 @@
+---
+kind: Deployment
+---
+kind: CronJob
