+++ infrastructure/helm/templates/built-in-workers.yaml
@@ -1,0 +1,2 @@
+---
+kind: Deployment
