+++ infrastructure/helm/templates/github.yaml
@@ -1,0 +1,6 @@
+---
+kind: Deployment
+---
+kind: Deployment
+---
+kind: CronJob
