+++ infrastructure/helm/templates/hooks.yaml
@@ -1,0 +1,8 @@
+---
+kind: Deployment
+---
+kind: Deployment
+---
+kind: Deployment
+---
+kind: CronJob
