+++ infrastructure/helm/templates/notify.yaml
@@ -1,0 +1,6 @@
+---
+kind: Deployment
+---
+kind: Deployment
+---
+kind: Deployment
