+++ infrastructure/helm/templates/purge-cache.yaml
@@ -1,0 +1,4 @@
+---
+kind: Deployment
+---
+kind: CronJob
