+++ infrastructure/helm/templates/queue.yaml
@@ -1,0 +1,26 @@
+---
+kind: Deployment
+---
+kind: Deployment
+---
+kind: Deployment
+---
+kind: Deployment
+---
+kind: CronJob
+---
+kind: CronJob
+---
+kind: CronJob
+---
+kind: CronJob
+---
+kind: CronJob
+---
+kind: CronJob
+---
+kind: CronJob
+---
+kind: CronJob
+---
+kind: CronJob
