+++ infrastructure/helm/templates/web-server.yaml
@@ -1,0 +1,2 @@
+---
+kind: Deployment
