+++ infrastructure/helm/templates/worker-manager.yaml
@@ -1,0 +1,4 @@
+---
+kind: Deployment
+---
+kind: Deployment
